test: refactor viewport controls and update snapshots
