+++ packages/core/components/ViewportControls/default-viewports.ts
@@ -1,0 +1,7 @@
+import { Viewports } from "../../types/Viewports";
+
+export const defaultViewports: Required<Viewports> = [
+  { width: 360, height: "auto", icon: "Smartphone", label: "Small" },
+  { width: 768, height: "auto", icon: "Tablet", label: "Medium" },
+  { width: 1280, height: "auto", icon: "Monitor", label: "Large" },
+];
