+++ packages/core/types/Config.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from "react";
 import { ReactNode } from "react";
 import { ItemSelector } from "../lib/get-item";
 import { DropZone } from "../components/DropZone";
-import { Viewport } from "../components/ViewportControls";
+import { Viewport } from "./Viewports";
 
 type WithPuckProps<Props> = Props & {
   id: string;
