+++ packages/core/types/Viewports.ts
@@ -1,0 +1,12 @@
+import { ReactNode } from "react";
+
+type iconTypes = "Smartphone" | "Monitor" | "Tablet";
+
+export type Viewport = {
+  width: number;
+  height?: number | "auto";
+  label?: string;
+  icon?: iconTypes | ReactNode;
+};
+
+export type Viewports = Viewport[];
